refactor: Move `process_description` to `tools.schemapi.utils`
Suspect this might have been used more dynamically in the past.
Currently doesn't make sense defining across 3 files
